+++ eras/conway/impl/src/Cardano/Ledger/Conway/Translation.hs
@@ -37,7 +37,6 @@ import Cardano.Ledger.Shelley.API (
   DState (..),
   EpochState (..),
   NewEpochState (..),
-  ShelleyGenesis,
   StrictMaybe (..),
  )
 import qualified Cardano.Ledger.Shelley.API as API
@@ -75,27 +74,6 @@ instance Crypto c => TranslateEra (ConwayEra c) NewEpochState where
         , nesRu = nesRu nes
         , nesPd = nesPd nes
         , stashedAVVMAddresses = ()
-        }
-
-instance Crypto c => TranslateEra (ConwayEra c) ShelleyGenesis where
-  translateEra ctxt genesis =
-    pure
-      API.ShelleyGenesis
-        { API.sgSystemStart = API.sgSystemStart genesis
-        , API.sgNetworkMagic = API.sgNetworkMagic genesis
-        , API.sgNetworkId = API.sgNetworkId genesis
-        , API.sgActiveSlotsCoeff = API.sgActiveSlotsCoeff genesis
-        , API.sgSecurityParam = API.sgSecurityParam genesis
-        , API.sgEpochLength = API.sgEpochLength genesis
-        , API.sgSlotsPerKESPeriod = API.sgSlotsPerKESPeriod genesis
-        , API.sgMaxKESEvolutions = API.sgMaxKESEvolutions genesis
-        , API.sgSlotLength = API.sgSlotLength genesis
-        , API.sgUpdateQuorum = API.sgUpdateQuorum genesis
-        , API.sgMaxLovelaceSupply = API.sgMaxLovelaceSupply genesis
-        , API.sgProtocolParams = translateEra' ctxt (API.sgProtocolParams genesis)
-        , API.sgGenDelegs = API.sgGenDelegs genesis
-        , API.sgInitialFunds = API.sgInitialFunds genesis
-        , API.sgStaking = API.sgStaking genesis
         }
 
 newtype Tx era = Tx {unTx :: Core.Tx era}
@@ -138,7 +116,7 @@ instance Crypto c => TranslateEra (ConwayEra c) EpochState where
         }
 
 instance Crypto c => TranslateEra (ConwayEra c) API.LedgerState where
-  translateEra ctxt@(ConwayGenesis newGenDelegs) ls =
+  translateEra ctxt@(ConwayGenesis _alonzoGenesis newGenDelegs) ls =
     pure
       API.LedgerState
         { API.lsUTxOState = translateEra' ctxt $ API.lsUTxOState ls