+++ eras/conway/test-suite/src/Test/Cardano/Ledger/Conway/Serialisation/Generators.hs
@@ -6,7 +6,9 @@ module Test.Cardano.Ledger.Conway.Serialisation.Generators () where
 
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
 import Cardano.Ledger.Crypto (Crypto)
+import Test.Cardano.Ledger.Alonzo.Serialisation.Generators ()
 import Test.Cardano.Ledger.Babbage.Serialisation.Generators ()
 import Test.QuickCheck (Arbitrary (..))
 
-deriving instance Crypto c => Arbitrary (ConwayGenesis c)
+instance Crypto c => Arbitrary (ConwayGenesis c) where
+  arbitrary = ConwayGenesis <$> arbitrary <*> arbitrary